\name{DCEtool}
\alias{DCEtool}
\title{Efficient and Accessible DCEs: DCEtool}
\description{Design, conduct, and analyze discrete choice experiments from a visual interface.}

\usage{
DCEtool()
}

\value{
Use the visual interface to generate, load and download designs and data bases.
}

\examples{
\dontrun{

  DCEtool()
  
  }
}